refactor: #252 토스트 컴포넌트 구조 리팩토링
addToast는 jsx로 사용하는 컴포넌트가 아닌
메세지와 옵션을 파라미터로 받아
특정 위치에 컴포넌트를 렌더해주는 함수입니다.

컨테이너 안에 있으면 컴포넌트와 혼동이 예상되었고,
다른 프로젝트에서도 사용할 수 있기 때문에 유틸로 분리할까 고민했습니다.

Component는 재사용가능한 UI를 의미하고
UI의 유틸 버전이라 판단하여

Component 폴더 안에분리되어있던 토스트와 관련된 파일을 모아
index.js에서 함수를 export하도록 구조를 리팩토링하였습니다.
